feat: restructure project directory and add page components for Home, About, Contact, and Login
