+++ 18_spa_vita/src/App.tsx
@@ -1,10 +1,10 @@
 
 import { useState } from 'react';
 import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
-import Home from './Home';
-import About from './About';
-import Contact from './Contact';
-import Login from './Login';
+import Home from './pages/Home';
+import About from './pages/About';
+import Contact from './pages/Contact';
+import Login from './pages/Login';
 
 function App() {
   const [loggedIn, setLoggedIn] = useState(false);
